fix(core): relocate duckdb_limits to core — keep runtime/doctor free of persistence imports
CI caught test_post_doctor_response_shape_matches_cli_json failing:
the function-local import of gispulse.persistence.duckdb_limits in
runtime/duckdb_engine.py triggered the persistence package init during
`gispulse doctor --json`, whose engine-backend registration logs
polluted stdout ahead of the JSON document. runtime/duckdb_engine
exists precisely so the doctor path never imports persistence.

Move the module to gispulse/core/duckdb_limits.py (it only depends on
core.config and core.logging) and update the nine call sites and the
test module accordingly. No behavioural change.
