+++ src/gispulse/core/io/geoparquet.py
@@ -184,7 +184,7 @@ def _read_via_duckdb(
         return _read_via_geopandas(path, bbox=bbox, columns=columns, **kwargs)
 
     try:
-        from gispulse.persistence.duckdb_limits import configure_duckdb_limits
+        from gispulse.core.duckdb_limits import configure_duckdb_limits
 
         conn = duckdb.connect(":memory:")
         configure_duckdb_limits(conn)
