+++ src/gispulse/persistence/bridge.py
@@ -60,7 +60,7 @@ class DuckDBPostGISBridge:
         Returns:
             Ready-to-use DuckDB connection.
         """
-        from gispulse.persistence.duckdb_limits import configure_duckdb_limits
+        from gispulse.core.duckdb_limits import configure_duckdb_limits
 
         conn = duckdb.connect(":memory:")
         configure_duckdb_limits(conn)
